Refactor initrd code out
